checkout-service/requirements.txt
flask
pika
pyjwt
